+++ CHANGELOG.md
@@ -5,5 +5,6 @@ FEATURES:
 * **New Resource:** `gigahost_dns_record`
 * **New Resource:** `gigahost_dns_zone`
 * **New Data Source:** `gigahost_account`
+* **New Data Source:** `gigahost_dns_records`
 * **New Data Source:** `gigahost_dns_zones`
 * **New Data Source:** `gigahost_ssh_keys`